+++ pwsh/AzGovVizParallel.ps1
@@ -280,7 +280,7 @@ Param
     $AzAPICallVersion = '1.1.11',
 
     [string]
-    $ProductVersion = 'v6_major_20220425_1',
+    $ProductVersion = 'v6_major_20220501_1',
 
     [string]
     $GithubRepository = 'aka.ms/AzGovViz',
@@ -8489,13 +8489,13 @@ extensions: [{ name: 'sort' }]
                 if ($roleAssignment.Scope -notlike 'this*') {
                     $rolesAssignedInheritedCount++
                 }
-                if ($roleAssignment.ObjectType -eq 'User') {
+                if ($roleAssignment.ObjectType -like 'User*') {
                     $rolesAssignedUser++
                 }
                 if ($roleAssignment.ObjectType -eq 'Group') {
                     $rolesAssignedGroup++
                 }
-                if ($roleAssignment.ObjectType -eq 'ServicePrincipal') {
+                if ($roleAssignment.ObjectType -like 'SP*') {
                     $rolesAssignedServicePrincipal++
                 }
                 if ($roleAssignment.ObjectType -eq 'Unknown') {
@@ -23928,23 +23928,22 @@ function dataCollectionRoleAssignmentsMG {
     )
 
     $addRowToTableDone = $false
-    #PIM MGRoleAssignmentSchedules
-    $currentTask = "getARMRoleAssignmentSchedules '$($scopeDisplayName)' ('$($scopeId)')"
-    $uri = "$($azAPICallConf['azAPIEndpointUrls'].ARM)/providers/Microsoft.Management/managementGroups/$($scopeId)/providers/Microsoft.Authorization/roleAssignmentSchedules?api-version=2020-10-01-preview"
+    #PIM MGRoleAssignmentScheduleInstances
+    $currentTask = "getARMRoleAssignmentScheduleInstances MG '$($scopeDisplayName)' ('$($scopeId)')"
+    $uri = "$($azAPICallConf['azAPIEndpointUrls'].ARM)/providers/Microsoft.Management/managementGroups/$($scopeId)/providers/Microsoft.Authorization/roleAssignmentScheduleInstances?api-version=2020-10-01"
     $method = 'GET'
-    $roleAssignmentSchedulesFromAPI = AzAPICall -AzAPICallConfiguration $azAPICallConf -uri $uri -method $method -currentTask $currentTask -caller 'CustomDataCollection'
+    $roleAssignmentScheduleInstancesFromAPI = AzAPICall -AzAPICallConfiguration $azAPICallConf -uri $uri -method $method -currentTask $currentTask -caller 'CustomDataCollection'
 
-    if ($roleAssignmentSchedulesFromAPI -eq 'ResourceNotOnboarded' -or $roleAssignmentSchedulesFromAPI -eq 'TenantNotOnboarded' -or $roleAssignmentSchedulesFromAPI -eq 'InvalidResourceType') {
+    if ($roleAssignmentScheduleInstancesFromAPI -eq 'ResourceNotOnboarded' -or $roleAssignmentScheduleInstancesFromAPI -eq 'TenantNotOnboarded' -or $roleAssignmentScheduleInstancesFromAPI -eq 'InvalidResourceType') {
         #Write-Host "Scope '$($scopeDisplayName)' ('$scopeId') not onboarded in PIM"
     }
     else {
-        $roleAssignmentSchedules = ($roleAssignmentSchedulesFromAPI.where( { -not [string]::IsNullOrEmpty($_.properties.roleAssignmentScheduleRequestId) }))
-        $roleAssignmentSchedulesCount = $roleAssignmentSchedules.Count
-        if ($roleAssignmentSchedulesCount -gt 0) {
+        $roleAssignmentScheduleInstances = ($roleAssignmentScheduleInstancesFromAPI.where( { ($_.properties.roleAssignmentScheduleId -replace '.*/') -ne ($_.properties.originRoleAssignmentId -replace '.*/') }))
+        $roleAssignmentScheduleInstancesCount = $roleAssignmentScheduleInstances.Count
+        if ($roleAssignmentScheduleInstancesCount -gt 0) {
             $htRoleAssignmentsPIM = @{}
-            foreach ($roleAssignmentSchedule in $roleAssignmentSchedules) {
-                $keyName = "$(($roleAssignmentSchedule.properties.scope).ToLower())-$(($roleAssignmentSchedule.properties.expandedProperties.principal.id).ToLower())-$(($roleAssignmentSchedule.properties.expandedProperties.roleDefinition.id).ToLower())"
-                $htRoleAssignmentsPIM.($keyName) = $roleAssignmentSchedule.properties
+            foreach ($roleAssignmentScheduleInstance in $roleAssignmentScheduleInstances) {
+                $htRoleAssignmentsPIM.($roleAssignmentScheduleInstance.properties.originRoleAssignmentId.tolower()) = $roleAssignmentScheduleInstance.properties
             }
         }
     }
@@ -23990,9 +23989,8 @@ function dataCollectionRoleAssignmentsMG {
     foreach ($L0mgmtGroupRoleAssignment in $L0mgmtGroupRoleAssignments) {
         $roleAssignmentId = ($L0mgmtGroupRoleAssignment.id).ToLower()
 
-        $keyName = "$(($L0mgmtGroupRoleAssignment.properties.scope).ToLower())-$(($L0mgmtGroupRoleAssignment.properties.principalId).ToLower())-$(($L0mgmtGroupRoleAssignment.properties.roleDefinitionId).ToLower())"
-        if ($htRoleAssignmentsPIM.($keyName)) {
-            $hlperPim = $htRoleAssignmentsPIM.($keyName)
+        if ($htRoleAssignmentsPIM.($roleAssignmentId)) {
+            $hlperPim = $htRoleAssignmentsPIM.($roleAssignmentId)
             $pim = 'true'
             $pimAssignmentType = $hlperPim.assignmentType
             $pimSlotStart = $($hlperPim.startDateTime)
@@ -24206,23 +24204,22 @@ function dataCollectionRoleAssignmentsSub {
 
     $script:htSubscriptionsRoleAssignmentLimit.($scopeId) = $roleAssignmentsUsage.roleAssignmentsLimit
 
-    #PIM SubscriptionRoleAssignmentSchedules
-    $currentTask = "Role assignment schedules API '$($scopeDisplayName)' ('$scopeId')"
-    $uri = "$($azAPICallConf['azAPIEndpointUrls'].ARM)/subscriptions/$($scopeId)/providers/Microsoft.Authorization/roleAssignmentSchedules?api-version=2020-10-01-preview"
+    #PIM SubscriptionRoleAssignmentScheduleInstances
+    $currentTask = "getARMRoleAssignmentScheduleInstances Sub '$($scopeDisplayName)' ('$scopeId')"
+    $uri = "$($azAPICallConf['azAPIEndpointUrls'].ARM)/subscriptions/$($scopeId)/providers/Microsoft.Authorization/roleAssignmentScheduleInstances?api-version=2020-10-01"
     $method = 'GET'
-    $roleAssignmentSchedulesFromAPI = AzAPICall -AzAPICallConfiguration $azAPICallConf -uri $uri -method $method -currentTask $currentTask -caller 'CustomDataCollection'
+    $roleAssignmentScheduleInstancesFromAPI = AzAPICall -AzAPICallConfiguration $azAPICallConf -uri $uri -method $method -currentTask $currentTask -caller 'CustomDataCollection'
 
-    if ($roleAssignmentSchedulesFromAPI -eq 'ResourceNotOnboarded' -or $roleAssignmentSchedulesFromAPI -eq 'TenantNotOnboarded' -or $roleAssignmentSchedulesFromAPI -eq 'InvalidResourceType') {
+    if ($roleAssignmentScheduleInstancesFromAPI -eq 'ResourceNotOnboarded' -or $roleAssignmentScheduleInstancesFromAPI -eq 'TenantNotOnboarded' -or $roleAssignmentScheduleInstancesFromAPI -eq 'InvalidResourceType') {
         #Write-Host "Scope '$($scopeDisplayName)' ('$scopeId') not onboarded in PIM"
     }
     else {
-        $roleAssignmentSchedules = ($roleAssignmentSchedulesFromAPI.where( { -not [string]::IsNullOrEmpty($_.properties.roleAssignmentScheduleRequestId) }))
-        $roleAssignmentSchedulesCount = $roleAssignmentSchedules.Count
-        if ($roleAssignmentSchedulesCount -gt 0) {
+        $roleAssignmentScheduleInstances = ($roleAssignmentScheduleInstancesFromAPI.where( { ($_.properties.roleAssignmentScheduleId -replace '.*/') -ne ($_.properties.originRoleAssignmentId -replace '.*/') }))
+        $roleAssignmentScheduleInstancesCount = $roleAssignmentScheduleInstances.Count
+        if ($roleAssignmentScheduleInstancesCount -gt 0) {
             $htRoleAssignmentsPIM = @{}
-            foreach ($roleAssignmentSchedule in $roleAssignmentSchedules) {
-                $keyName = "$(($roleAssignmentSchedule.properties.scope).ToLower())-$(($roleAssignmentSchedule.properties.expandedProperties.principal.id).ToLower())-$(($roleAssignmentSchedule.properties.expandedProperties.roleDefinition.id).ToLower())"
-                $htRoleAssignmentsPIM.($keyName) = $roleAssignmentSchedule.properties
+            foreach ($roleAssignmentScheduleInstance in $roleAssignmentScheduleInstances) {
+                $htRoleAssignmentsPIM.($roleAssignmentScheduleInstance.properties.originRoleAssignmentId.tolower()) = $roleAssignmentScheduleInstance.properties
             }
         }
     }
@@ -24359,9 +24356,8 @@ function dataCollectionRoleAssignmentsSub {
             $roleAssignmentScopeRes = $roleAssignmentScopeSplit[8]
         }
 
-        $keyName = "$(($L1mgmtGroupSubRoleAssignment.properties.scope).ToLower())-$(($L1mgmtGroupSubRoleAssignment.properties.principalId).ToLower())-$(($L1mgmtGroupSubRoleAssignment.properties.roleDefinitionId).ToLower())"
-        if ($htRoleAssignmentsPIM.($keyName)) {
-            $hlperPim = $htRoleAssignmentsPIM.($keyName)
+        if ($htRoleAssignmentsPIM.($roleAssignmentId)) {
+            $hlperPim = $htRoleAssignmentsPIM.($roleAssignmentId)
             $pim = 'true'
             $pimAssignmentType = $hlperPim.assignmentType
             $pimSlotStart = $($hlperPim.startDateTime)